# openapi.yaml
openapi: 3.0.0
info:
  title: BankWiz API
  version: 0.1.0
  description: An API for accessing banking functionalities provided by BankWiz
servers:
  - url: 'http://localhost:8080/api'
...
paths:
  /auth/signup:
    post:
      tags:
        - Auth
      summary: Create a new user account
      operationId: createUserAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserAccountRequest'
      responses:
        '201':
          description: User account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Invalid request. Please check the provided data.
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login to an existing user account
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Incorrect request
        '401':
          description: Unauthorized
...
  /bankAccount:
    post:
      tags:
        - Bank account
      summary: Add an account
      operationId: addAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankAccountCreationRequest'
      responses:
        '200':
          description: Account added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccountDTO'
        '400':
          description: Invalid request. Please check the provided data.
  /bankAccount/bankAccounts:
    get:
      tags:
        - Bank account
      summary: Get my bank accounts
      operationId: getBankAccounts
      responses:
        '200':
          description: Successfully retrieved the list of bank accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BankAccountGroupDTO'
        '400':
          description: Invalid request. Please check the provided data.
  '/bankAccount/{bankAccountId}':
    get:
      tags:
        - Bank account
      parameters:
        - name: bankAccountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      summary: Get an account
      operationId: getAccount
      responses:
        '200':
          description: Successfully retrieved account information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccountDTO'
        '400':
          description: Invalid request. Please check the provided data.
    put:
      tags:
        - Bank account
      parameters:
        - name: bankAccountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      summary: Update account information
      operationId: updateAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankAccountUpdateRequest'
      responses:
        '200':
          description: Bank account information updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccountDTO'
        '400':
          description: Invalid request. Please check the provided data.
    delete:
      tags:
        - Bank account
      parameters:
        - name: bankAccountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      summary: Delete an account
      operationId: deleteAccount
      responses:
        '200':
          description: Bank account deleted successfully
        '400':
          description: Invalid request. Please check the provided data.
...
components:
  schemas:
    CreateUserAccountRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - firstName
        - lastName
        - email
        - password
...
    UserLoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - email
        - password
...
    BankAccountCreationRequest:
      type: object
      properties:
        groupId:
          type: integer
        accountName:
          type: string
        baseAmountDecimal:
          type: integer
      required:
        - groupId
        - accountName
        - baseAmountDecimal
    BankAccountUpdateRequest:
      type: object
      properties:
        accountName:
          type: string
        baseAmountDecimal:
          type: integer
...
    UserDTO:
      type: object
      properties:
        userId:
          type: integer
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: johndoe@example.com
      required:
        - userId
        - firstName
        - lastName
        - email
...
    GroupIndexDTO:
      type: object
      properties:
        groupId:
          type: integer
        groupName:
          type: string
      required:
        - groupId
        - groupName
    BankAccountIndexDTO:
      type: object
      properties:
        accountId:
          type: integer
        accountName:
          type: string
        baseAmountDecimal:
          type: integer
      required:
        - accountId
        - accountName
        - baseAmountDecimal
    BankAccountDTO:
      type: object
      properties:
        groupIndexDTO:
          $ref: '#/components/schemas/GroupIndexDTO'
        accountIndexDTO:
          $ref: '#/components/schemas/BankAccountIndexDTO'
      required:
        - groupIndexDTO
        - accountIndexDTO
    BankAccountGroupDTO:
      type: object
      properties:
        groupIndexDTO:
          $ref: '#/components/schemas/GroupIndexDTO'
        accountIndexDTOs:
          type: array
          items:
            $ref: '#/components/schemas/BankAccountIndexDTO'
      required:
        - groupIndexDTO
        - accountIndexDTOs
...